{"schema":"decodex/commit/1","summary":"Batch modularize oversized validation and acceptance helpers","authority":"manual"}
