+++ packages/elf-service/tests/acceptance/trace_admin_observability.rs
@@ -2,9 +2,13 @@ mod helpers;
 
 pub(crate) use helpers::{
 	PROJECT_ID, TENANT_ID, TraceAdminObservabilityFixture, VisibilityTraceFixtureIds,
-	assert_trace_admin_visibility_cross_scope, insert_trace, insert_trace_candidate,
-	insert_trace_item, insert_trace_stage, insert_trace_stage_item,
-	seed_visibility_and_recent_list_traces, setup_service, trace_recent_list_page,
+	assertions::assert_trace_admin_visibility_cross_scope,
+	inserts::{
+		insert_trace, insert_trace_candidate, insert_trace_item, insert_trace_stage,
+		insert_trace_stage_item,
+	},
+	seed::{seed_visibility_and_recent_list_traces, trace_recent_list_page},
+	setup::setup_service,
 };
 
 use time::OffsetDateTime;
