+++ packages/elf-service/tests/acceptance/trace_admin_observability/helpers/setup.rs
@@ -1,0 +1,44 @@
+use std::sync::{Arc, atomic::AtomicUsize};
+
+use crate::acceptance::{
+	self, SpyExtractor, StubEmbedding, StubRerank,
+	trace_admin_observability::helpers::TraceAdminObservabilityFixture,
+};
+use elf_service::Providers;
+
+pub(crate) async fn setup_service(test_name: &str) -> Option<TraceAdminObservabilityFixture> {
+	let Some(test_db) = acceptance::test_db().await else {
+		eprintln!("Skipping {test_name}; set ELF_PG_DSN to run this test.");
+
+		return None;
+	};
+	let Some(qdrant_url) = acceptance::test_qdrant_url() else {
+		eprintln!("Skipping {test_name}; set ELF_QDRANT_URL to run this test.");
+
+		return None;
+	};
+	let collection = test_db.collection_name("elf_acceptance");
+	let docs_collection = test_db.collection_name("elf_acceptance_docs");
+	let cfg = acceptance::test_config(
+		test_db.dsn().to_string(),
+		qdrant_url,
+		4_096,
+		collection,
+		docs_collection,
+	);
+	let extractor = SpyExtractor {
+		calls: Arc::new(AtomicUsize::new(0)),
+		payload: serde_json::json!({ "notes": [] }),
+	};
+	let providers = Providers::new(
+		Arc::new(StubEmbedding { vector_dim: 4_096 }),
+		Arc::new(StubRerank),
+		Arc::new(extractor),
+	);
+	let service =
+		acceptance::build_service(cfg, providers).await.expect("Failed to build service.");
+
+	acceptance::reset_db(&service.db.pool).await.expect("Failed to reset test database.");
+
+	Some(TraceAdminObservabilityFixture { service, test_db })
+}
